dgsm/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "dgsm"
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "aioconsole",
    "colorama",
    "netifaces",
    "psutil",
    "PyYAML",
    "upnpclient",
]